started a loader to create entities from file and changed int to float for the movements

+++ src/main/scala/simulation/MovementState.scala
@@ -3,15 +3,15 @@ package simulation
 import org.lwjgl.glfw.GLFW.{GLFW_KEY_A, GLFW_KEY_D, GLFW_KEY_S, GLFW_KEY_W, GLFW_RELEASE}
 
 object d extends Enumeration{
-  val LEFT: (Int, Int) = (-1,0)
-  val UP: (Int, Int) = (0, 1)
-  val DOWN: (Int, Int) = (0, -1)
-  val RIGHT: (Int, Int) = (1, 0)
+  val LEFT: (Float, Float) = (-1,0)
+  val UP: (Float, Float) = (0, 1)
+  val DOWN: (Float, Float) = (0, -1)
+  val RIGHT: (Float, Float) = (1, 0)
 }
 
-sealed abstract class MovementState(dir : (Int, Int)){
+sealed abstract class MovementState(dir : (Float, Float)){
   def inputChanged(stateMachine: MovementStateMachine, key : Int)
-  def getNewDir(newDir : (Int, Int)): (Int, Int) = (dir._1 + newDir._1, dir._2 + newDir._2)
+  def getNewDir(newDir : (Float, Float)): (Float, Float) = (dir._1 + newDir._1, dir._2 + newDir._2)
 }
 
 case class Standing() extends MovementState((0,0)) {
@@ -26,7 +26,7 @@ case class Standing() extends MovementState((0,0)) {
   override def toString: String = "Standing"
 }
 
-case class Walking(dir : (Int, Int)) extends MovementState(dir) {
+case class Walking(dir : (Float, Float)) extends MovementState(dir) {
   override def inputChanged(stateMachine: MovementStateMachine, key : Int): Unit = key match {
     case GLFW_KEY_A => stateMachine.pushState(Walking(getNewDir(d.LEFT)))
     case GLFW_KEY_W => stateMachine.pushState(Walking(getNewDir(d.UP)))
